agent/requirements.txt
web3==7.6.0
eth-account==0.13.4
python-dotenv==1.0.1
aiohttp==3.11.11
asyncio==3.4.3
httpx==0.28.1
pydantic==2.10.6
loguru==0.7.3
schedule==1.2.2
